Move fixer test to fix package
