+++ command/fix_test.go
@@ -1,7 +1,6 @@
 package command
 
 import (
-	"github.com/hashicorp/packer/fix"
 	"path/filepath"
 	"strings"
 	"testing"
@@ -60,29 +59,5 @@ func TestFix_invalidTemplateDisableValidation(t *testing.T) {
 	}
 	if code := c.Run(args); code != 0 {
 		fatalCommand(t, c.Meta)
-	}
-}
-
-func TestFix_allFixersEnabled(t *testing.T) {
-	f := fix.Fixers
-	o := fix.FixerOrder
-
-	if len(f) != len(o) {
-		t.Fatalf("Fixers length (%d) does not match FixerOrder length (%d)", len(f), len(o))
-	}
-
-	for fixer, _ := range f {
-		found := false
-
-		for _, orderedFixer := range o {
-			if orderedFixer == fixer {
-				found = true
-				break
-			}
-		}
-
-		if !found {
-			t.Fatalf("Did not find Fixer %s in FixerOrder", fixer)
-		}
 	}
 }
